+++ src/schemas/memberRole.ts
@@ -17,6 +17,8 @@ export const typeDefs = gql`
     updateMemberRole(id: Int!, url: URL!, label: String!, displayOrder: Int!, description: String): MemberRole
     "Delete the member role"
     removeMemberRole(id: Int!): MemberRole
+    "Mark the member role as the default role"
+    setDefaultMemberRole(id: Int!): MemberRole
   }
 
   type MemberRole {
@@ -41,6 +43,8 @@ export const typeDefs = gql`
     uri: String!
     "A longer description of the member role useful for tooltips"
     description: String
+    "Whether this is the default role"
+    isDefault: Boolean
   }
 
   "A collection of errors related to the member role"